some minor updates

+++ scripts/analysis/plot_folded_dssp.py
@@ -26,17 +26,21 @@ if __name__ == '__main__':
     import matplotlib.pyplot as plt
     import seaborn as sns
     from tqdm import tqdm
+    import os
     mpl.rcParams.update({
         'font.size': 18,
-        'figure.figsize': (9, 6)
+        # 'figure.figsize': (9, 6)
     })
 
-    results = []
-    for pdb in tqdm(glob.glob("precise_samples_folded/*_sc.pdb")):
-        results.append(compute_dssp(pdb))
+    if os.path.exists("dssp_per_precise_sample.csv"):
+        df = pd.read_csv("dssp_per_precise_sample.csv")
+    else:
+        results = []
+        for pdb in tqdm(glob.glob("precise_samples_folded/*_sc.pdb")):
+            results.append(compute_dssp(pdb))
 
-    df = pd.DataFrame(results)
-    df.to_csv("dssp_per_precise_sample.csv")
+        df = pd.DataFrame(results)
+        df.to_csv("dssp_per_precise_sample.csv")
     # df = pd.read_csv("dssp_per_precise_sample.csv")
     df['frac_helix'] = df['helix_percent']
     df['frac_strand'] = df['strand_percent']
@@ -47,7 +51,8 @@ if __name__ == '__main__':
     else:
         discrete = False
         colors = None
-    jointplot = sns.jointplot(df, x="frac_helix", y="frac_strand", hue="seq_len", palette=colors, height=9, s=100)
+    # jointplot = sns.jointplot(df, x="frac_helix", y="frac_strand", hue="seq_len", palette=colors, height=9, s=100)
+    jointplot = sns.jointplot(df, x="frac_helix", y="frac_strand", hue="seq_len", palette=colors, height=6.6, ratio=6, s=100)
     # plt.suptitle("Secondary structure content of\nsequence-structure consistent samples")
 
     ax = plt.gca()
